Move home activity to fragment
