+++ app/src/main/AndroidManifest.xml
@@ -40,7 +40,6 @@
             <category android:name="android.intent.category.LAUNCHER" />
         </intent-filter>
         </activity>
-        <activity android:name=".activities.HomeActivity" />
         <activity android:name=".activities.LoginActivity" />
         <activity android:name=".activities.TripActivity" />
         <activity
