--- app/src/main/java/turtler/voyageur/activities/HomeActivity.java
@@ -1,14 +1,0 @@
-package turtler.voyageur.activities;
-
-import android.os.Bundle;
-
-/**
- * Created by carolinewong on 9/3/16.
- */
-public class HomeActivity extends BaseActivity {
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-}
